+++ packages/adminui-pub/_src/app/view/EntityGridPanel.js
@@ -960,7 +960,7 @@ Ext.define('UB.view.EntityGridPanel', {
       popupMenuItems.push(me.actions[actions.audit])
     }
 
-    if (me.hasHardSecurityMixin) {
+    if (me.hasHardSecurityMixin && !me.domainEntity.mixins.aclRls.sameAs) {
       popupMenuItems.push('-')
       popupMenuItems.push(me.actions[actions.accessRight])
     }